feat(light): add light (#18)

+++ engine/src/classes/modelConfig/modelConfig.hpp
@@ -14,54 +14,55 @@ private:
   string color;
 
   // default color attributes
-  tuple<int, int, int> diffuse = make_tuple(200, 200, 200);
-  tuple<int, int, int> ambient = make_tuple(50, 50, 50);
-  tuple<int, int, int> specular = make_tuple(0, 0, 0);
-  tuple<int, int, int> emissive = make_tuple(0, 0, 0);
-  int shininess = 0;
+  tuple<float, float, float> diffuse = make_tuple(200, 200, 200);
+  tuple<float, float, float> ambient = make_tuple(50, 50, 50);
+  tuple<float, float, float> specular = make_tuple(0, 0, 0);
+  tuple<float, float, float> emissive = make_tuple(0, 0, 0);
+  float shininess = 0;
 
 public:
   ModelConfig();
 
   ModelConfig(string texture_file_path, string color,
-              tuple<int, int, int> diffuse, tuple<int, int, int> ambient,
-              tuple<int, int, int> specular, tuple<int, int, int> emissive,
-              int shininess);
+              tuple<float, float, float> diffuse,
+              tuple<float, float, float> ambient,
+              tuple<float, float, float> specular,
+              tuple<float, float, float> emissive, float shininess);
 
   string get_texture_file_path();
 
   string get_color();
 
-  tuple<int, int, int> get_diffuse();
+  tuple<float, float, float> get_diffuse();
 
-  tuple<int, int, int> get_ambient();
+  tuple<float, float, float> get_ambient();
 
-  tuple<int, int, int> get_specular();
+  tuple<float, float, float> get_specular();
 
-  tuple<int, int, int> get_emissive();
+  tuple<float, float, float> get_emissive();
 
-  int get_shininess();
+  float get_shininess();
 
   void set_texture_file_path(string path);
 
   void set_color(string color);
 
-  void set_diffuse(tuple<int, int, int> diffuse);
+  void set_diffuse(tuple<float, float, float> diffuse);
 
-  void set_diffuse(int r, int g, int b);
+  void set_diffuse(float r, float g, float b);
 
-  void set_ambient(tuple<int, int, int> ambient);
+  void set_ambient(tuple<float, float, float> ambient);
 
-  void set_ambient(int r, int g, int b);
+  void set_ambient(float r, float g, float b);
 
-  void set_specular(tuple<int, int, int> specular);
+  void set_specular(tuple<float, float, float> specular);
 
-  void set_specular(int r, int g, int b);
+  void set_specular(float r, float g, float b);
 
-  void set_emissive(tuple<int, int, int> emissive);
+  void set_emissive(tuple<float, float, float> emissive);
 
-  void set_emissive(int r, int g, int b);
+  void set_emissive(float r, float g, float b);
 
-  void set_shininess(int shininess);
+  void set_shininess(float shininess);
 };
 #endif
